Update Activity 1.py

+++ Assignment 4/Activity 1.py
@@ -1,8 +1,8 @@
 # Create a program to prompt the user for hours and rate per hour and then calculate and display their weekly, monthly, and annual gross pay (hours * rate). Base monthly and annual calculations on 12 months per year and 52 weeks per year.
 print(" How many weekly hours did you work ?")
-hours = int(input())
+hours = float(input())
 print(" What is your rate foy pay ?")
-rate = int(input())
+rate = float(input())
 weekly = rate * hours
 monthly = weekly * 30
 yearly = weekly * 365
